2D_ROI/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
blinker==1.5
Bottleneck==1.3.5
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.32
docker-pycreds==0.4.0
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
h5py==3.7.0
idna==3.4
imageio==2.22.4
importlib-metadata==4.11.4
joblib==1.1.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
munkres==1.1.4
nibabel==4.0.2
numexpr==2.8.3
numpy==1.21.5
oauthlib==3.2.2
opencv-python==4.6.0
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
Pillow==9.2.0
pip==22.3.1
ply==3.11
promise==2.3
protobuf==4.21.8
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycocotools==2.0.4
pycparser==2.21
pyDeprecate==0.3.2
pydicom==2.3.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytz==2022.1
pyu2f==0.1.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentry-sdk==1.10.1
setproctitle==1.3.2
setuptools==65.5.1
shortuuid==1.0.10
sip==6.6.2
six==1.16.0
smmap==5.0.0
tensorboard==2.11.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==2.2.0
toml==0.10.2
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tornado==6.1
tqdm==4.64.1
typing_extensions==4.4.0
unicodedata2==14.0.0
urllib3==1.26.11
wandb==0.13.5
Werkzeug==2.2.2
wheel==0.38.3
yarl==1.7.2
zipp==3.10.0
